Initial Commit
A revised version of the NFL Team Picker Recoded in React.

Popped all the names in Firebase too.

Right now I'm trying to get the images to appear. Webpack means it's not as simple to add the local image files. Found a slice of code that's seem to have done the job, but im gonna need to do some tinkering to get the correct images up, (or indeed, any images at all).

TO DO:

- Get images showing up on the GUI
- Display the Images in a spinner carousel. Get the ideal 'scroll to chosen image' functionality going so that the spinner can work as intended
- Get Spinner Working. (i.e. get the carousel to spin to a randomly selected slide.)