Final Commit
All working!

Pretty much just hardcoded the remaining images. I got what I needed from this project. Which was to:

- Learn Firebase uploads
- Learn the Antd framework system

AntD is a bit slow to load, but I'd be interested in learning more about it in the future. Had a few issues with formatting everything too, but other than that. I'm happy I got this all working!